requirements.txt
Django>=4.2
mysqlclient>=2.2.0
PyMySQL>=1.1.0
psycopg2-binary>=2.9.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
python-dotenv>=1.0.0
djangorestframework-simplejwt>=5.2.2
pandas>=2.1.4
geopandas>=0.14.1
pyogrio>=0.6.0
shapely>=2.0.2
folium>=0.15.0
reportlab>=4.0.0
Pillow>=10.0.0
selenium>=4.15.0 